[refactor] group 패키지 구조 변경
- 패키지에 GroupService를 추가하고 데이터 클래스를 entity 패키지로 묶음.
