+++ mobile/app/src/main/java/com/lacuc/pets/data/group/GroupDataSource.kt
@@ -1,5 +1,7 @@
 package com.lacuc.pets.data.group
 
+import com.lacuc.pets.data.group.entity.Group
+
 interface GroupDataSource {
     fun loadGroup(email: String): List<Group>
