+++ mobile/app/src/main/java/com/lacuc/pets/domain/group/AddGroupUseCase.kt
@@ -1,7 +1,7 @@
 package com.lacuc.pets.domain.group
 
-import com.lacuc.pets.data.group.Group
 import com.lacuc.pets.data.group.GroupRepository
+import com.lacuc.pets.data.group.entity.Group
 import javax.inject.Inject
 
 class AddGroupUseCase @Inject constructor(private val repository: GroupRepository) {
