+++ mobile/app/src/test/java/com/lacuc/pets/domain/group/AddGroupUseCaseTest.kt
@@ -2,8 +2,8 @@ package com.lacuc.pets.domain.group
 
 import com.lacuc.pets.data.group.DefaultGroupRepository
 import com.lacuc.pets.data.group.FakeGroupRemoteDataSource
-import com.lacuc.pets.data.group.Group
 import com.lacuc.pets.data.group.GroupRepository
+import com.lacuc.pets.data.group.entity.Group
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
